refactor(actions): TMDB actions를 도메인 중심 구조로 재구성
- movie.ts를 tmdb/movie.ts로 이동하여 도메인별 응집도 향상
- ActionResult 및 파라미터 스키마를 types.ts로 분리
- Utils 함수들을 tmdb/utils/ 폴더로 분리
  - config.ts: API 기본 URL 상수
  - auth.ts: 인증 헤더 생성 함수
  - error.ts: TMDB 에러 처리 함수
  - index.ts: Utils 통합 export
- Action 함수는 비즈니스 로직에만 집중하도록 개선
- 향후 다른 API 도메인 추가 시 확장 용이
