+++ movie/src/actions/tmdb/utils/config.ts
@@ -1,0 +1,5 @@
+/**
+ * TMDB API Base URL
+ */
+export const TMDB_API_BASE_URL = "https://api.themoviedb.org/3"
+
